move filename to source_base
