+++ source/source_io/module_hs/write_vxc_lip.hpp
@@ -9,7 +9,7 @@
 #include "source_cell/klist.h"
 #include "source_estate/module_pot/potential_new.h"
 #include "source_io/module_hs/write_HS.h"
-#include "source_io/module_output/filename.h" // use filename_output function
+#include "source_base/module_out/filename.h" // use filename_output function
 #include <type_traits>
 
 namespace ModuleIO
