+++ source/source_io/test/support/wf_nao.txt
@@ -1,0 +1,14 @@
+3 (number of bands)
+3 (number of orbitals)
+1 (band)
+-7.5199188050468579458396334e-01 (Ry)
+2.5000000000000000000000000e-01 (Occupations)
+5.3759239842e-01 3.7570058235e-02 1.5292654624e-02
+2 (band)
+3.1482195194888534794941393e-01 (Ry)
+0.0000000000000000000000000e+00 (Occupations)
+1.8587183851e+00 5.9595655928e-01 5.3528900615e-02 
+3 (band)
+7.4141254894954844445464914e-01 (Ry)
+0.0000000000000000000000000e+00 (Occupations)
+3.5389238463e-02 5.6401881891e-01 2.3262992046e-02
