+++ source/source_psi/psi_init_file.cpp
@@ -3,7 +3,7 @@
 #include "source_base/timer.h"
 #include "source_cell/klist.h"
 #include "source_io/module_wf/read_wfc_pw.h"
-#include "source_io/module_output/filename.h"
+#include "source_base/module_out/filename.h"
 #include "source_io/module_parameter/parameter.h"
 
 template <typename T>
