+++ source/source_psi/test/CMakeLists.txt
@@ -19,7 +19,6 @@ AddTest(
         ../../source_io/module_output/orb_io.cpp
         ../../source_io/module_output/write_pao.cpp
         ../../source_io/module_wf/read_wfc_pw.cpp
-        ../../source_io/module_output/filename.cpp
 )
 endif()
